+++ CHANGELOG.md
@@ -26,7 +26,7 @@
 
 ## Fixed Issues
 
--
+- [core] Fix verification key retrieval to match the given key id (`kid`).
 
 
 # 1.48.1